backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
openai==1.3.9
langchain==0.1.4
langchain-community==0.0.13
faiss-cpu==1.7.4
tiktoken==0.5.2
langchain-openai==0.0.5
deep-translator==1.11.4
PyPDF2==3.17.0
pdfplumber==0.10.3
requests==2.31.0
